rejection window-alert enabled for multiple selects
